+++ packages/core/test/actor.test.ts
@@ -30,10 +30,7 @@ import {
   stopChild
 } from '../src/index.ts';
 import { setup } from '../src/setup.ts';
-
-function sleep(ms: number) {
-  return new Promise((res) => setTimeout(res, ms));
-}
+import { sleep } from '@xstate-repo/jest-utils';
 
 describe('spawning machines', () => {
   const context = {
@@ -391,6 +388,44 @@ describe('spawning callbacks', () => {
 
     callbackService.start();
     callbackService.send({ type: 'START_CB' });
+  });
+
+  it('should not deliver events sent to the parent after the callback actor gets stopped', () => {
+    const spy = jest.fn();
+
+    let sendToParent: () => void;
+
+    const machine = createMachine({
+      initial: 'a',
+      states: {
+        a: {
+          invoke: {
+            src: fromCallback(({ sendBack }) => {
+              sendToParent = () =>
+                sendBack({
+                  type: 'FROM_CALLBACK'
+                });
+            })
+          },
+          on: {
+            NEXT: 'b'
+          }
+        },
+        b: {}
+      },
+      on: {
+        FROM_CALLBACK: {
+          actions: spy
+        }
+      }
+    });
+
+    const actorRef = createActor(machine).start();
+    actorRef.send({ type: 'NEXT' });
+
+    sendToParent!();
+
+    expect(spy).not.toHaveBeenCalled();
   });
 });