+++ packages/xstate-svelte/test/UseActor.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   export let persistedState: AnyMachineSnapshot | undefined = undefined;
 
-  import { useMachine } from '@xstate/svelte';
+  import { useActor } from '@xstate/svelte';
   import { fetchMachine } from './fetchMachine.ts';
   import type { AnyMachineSnapshot } from 'xstate';
   import { fromPromise } from 'xstate/actors';
@@ -9,23 +9,27 @@
   const onFetch = () =>
     new Promise<string>((res) => setTimeout(() => res('some data'), 50));
 
-  const { state, send } = useMachine(fetchMachine, {
-    state: persistedState,
-    actors: {
-      fetchData: fromPromise(onFetch)
+  const { snapshot, send } = useActor(
+    fetchMachine.provide({
+      actors: {
+        fetchData: fromPromise(onFetch)
+      }
+    }),
+    {
+      state: persistedState
     }
-  });
+  );
 </script>
 
 <div>
-  {#if $state.matches('idle')}
+  {#if $snapshot.matches('idle')}
     <button on:click={() => send({ type: 'FETCH' })}>Fetch</button>
-  {:else if $state.matches('loading')}
+  {:else if $snapshot.matches('loading')}
     <div>Loading...</div>
-  {:else if $state.matches('success')}
+  {:else if $snapshot.matches('success')}
     <div>
       Success! Data:
-      <div data-testid="data">{$state.context.data}</div>
+      <div data-testid="data">{$snapshot.context.data}</div>
     </div>
   {/if}
 </div>